Double check the file exists before deleting in File cache adapter.

+++ storage/cache/adapter/File.php
@@ -105,7 +105,7 @@ class File extends \lithium\core\Object {
 			$expiry = $matches[1];
 
 			if ($expiry < time()) {
-				unlink($path);
+				file_exists($path) && unlink($path);
 				return false;
 			}
 			return preg_replace('/^\{\:expiry\:\d+\}\\n/', '', $data, 1);
@@ -127,7 +127,7 @@ class File extends \lithium\core\Object {
 			$file = new SplFileInfo($path);
 
 			if ($file->isFile() && $file->isReadable()) {
-				return unlink($path);
+				return file_exists($path) && unlink($path);
 			}
 			return false;
 		};
